Fixed reinterpret_cast warning on Windows when casting a 32-bit number to a 64-bit address

+++ OpenNeighborhood/src/Elements/AddXboxButton.cpp
@@ -18,7 +18,7 @@ void AddXboxButton::OnRender()
 {
     auto texture = TextureManager::GetTexture(m_TextureName);
 
-    if (ImGui::ImageButtonWithText(reinterpret_cast<void *>(texture->GetTextureID()), ImVec2(texture->GetWidth(), texture->GetHeight()), ImVec2(m_Width, m_Height), m_Label.c_str(), ImVec2(m_Padding, m_Padding)))
+    if (ImGui::ImageButtonWithText(reinterpret_cast<void *>(static_cast<intptr_t>(texture->GetTextureID())), ImVec2(texture->GetWidth(), texture->GetHeight()), ImVec2(m_Width, m_Height), m_Label.c_str(), ImVec2(m_Padding, m_Padding)))
         if (ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left))
             OnClick();
